backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.3
sqlalchemy==2.0.32
psycopg[binary]==3.2.1; python_version < "3.13"
psycopg==3.2.1; python_version >= "3.13"
pydantic==2.8.2
pydantic-settings==2.2.1
httpx==0.27.0
python-dateutil==2.9.0.post0
alembic==1.13.2
loguru==0.7.2
redis==5.0.7
openai>=1.42.0
pyyaml==6.0.2
